:sparkles: :: 애니메이션 적용

+++ SwiftUI/Summary/TodoList/TodoList/Views/ListView.swift
@@ -13,20 +13,33 @@ struct ListView: View {
     
     var body: some View {
         
-        List {
-            ForEach(listViewModel.items,id: \.self.id){ item in
-                ListRowView(item:item)
-                    .onTapGesture {
-                        withAnimation(.linear ) {
-                            listViewModel.updateItem(item: item)
-                        }
-                    }
+        
+        ZStack{
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition(.opacity.animation(.easeIn))
                 
+            } else {
+                List {
+                    ForEach(listViewModel.items,id: \.self.id){ item in
+                        ListRowView(item:item)
+                            .onTapGesture {
+                                withAnimation(.linear ) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
+                        
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle()) // 밑줄 구분자 있는 스타일
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
+            
         }
-        .listStyle(PlainListStyle()) // 밑줄 구분자 있는 스타일
+        
+        
+     
         .navigationTitle("Todo List ✏️")
         .toolbar {
             ToolbarItem(placement:.navigationBarLeading) {
